Add dynamic routing for proposal form steps
